+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - `Data` instances for the various data types.
 - Simple conversion types between the main version types.
+- Compile-time constructors via Template Haskell, like `versioningQ`.
 
 ## 6.0.2 (2023-10-12)